+++ src/App.jsx
@@ -1,26 +1,10 @@
-import { useState, useMemo } from "react";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 import MasterLayout from "./layout/MasterLayout";
 import HomePage from "./pages/Home";
 import AboutPage from "./pages/About";
 import ErrorPage from "./pages/Error";
-
-import { ThemeProvider, createTheme } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
 import "./sass/app.scss";
-
-const lightTheme = createTheme({
-  palette: {
-    mode: "light",
-  },
-});
-
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-  },
-});
 
 const router = createBrowserRouter([
   {
@@ -35,21 +19,21 @@ const router = createBrowserRouter([
 ]);
 
 function App() {
-  const [currentTheme, setCurrentTheme] = useState("light");
+  //const [currentTheme, setCurrentTheme] = useState("light");
 
-  const toggleTheme = () => {
-    setCurrentTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  };
+  // const toggleTheme = () => {
+  //   console.log('toggling theme');
+  //   setCurrentTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  // };
 
-  const theme = useMemo(() => {
-    return currentTheme === "light" ? lightTheme : darkTheme;
-  }, [currentTheme]);
+  // const theme = useMemo(() => {
+  //   return currentTheme === "light" ? lightTheme : darkTheme;
+  // }, [currentTheme]);
 
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <RouterProvider router={router} />
-    </ThemeProvider>
+    <>      
+      <RouterProvider router={router}  />
+    </>
   );
 }
