+++ src/layout/MasterLayout.jsx
@@ -1,13 +1,48 @@
+import { useState, useMemo } from "react";
 import { Outlet } from "react-router-dom";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
 import Header from "./sections/Header";
 
+// const theme = createTheme({
+//   palette: {
+//     mode: "light",
+//   },
+// });
+
+const lightTheme = createTheme({
+  palette: {
+    mode: "light",
+  },
+});
+
+const darkTheme = createTheme({
+  palette: {
+    mode: "dark",
+  },
+});
+
 function MasterLayout() {
+  const [currentTheme, setCurrentTheme] = useState("light");
+
+  const toggleTheme = () => {
+    console.log("toggling theme");
+    setCurrentTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  };
+
+  const theme = useMemo(() => {
+    return currentTheme === "light" ? lightTheme : darkTheme;
+  }, [currentTheme]);
+
   return (
     <>
-      <Header />
-      <Outlet />
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <Header toggleTheme={toggleTheme} />
+        <Outlet />
+      </ThemeProvider>
     </>
-  )
+  );
 }
 
 export default MasterLayout;
